Y3S1/CP/byteshare/backend/requirements.txt
click==6.7
Flask==1.1.2
Flask-HTTPAuth==4.0.0
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
PyJWT==1.7.1
SQLAlchemy==1.3.6
Werkzeug==1.0.1
